requirements.txt
python-dotenv==0.21.0
selenium==4.5.0
schedule==1.1.0
